requirements.txt
flask_cors 
flask
tensorflow
Pillow